requirements.txt
asgiref==3.3.2
Django==3.2.25
django-appconf==1.0.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-select2==8.2.0
pillow==10.3.0
pytz==2020.1
sqlparse==0.4.4
